// VersionControlBackend/models/user.model.ts
import { Schema, model } from 'mongoose';

/**
 * user schema
 * 
 */
const UserSchema: Schema = new Schema(
    {
        // _id is auto-generated by mongodb
        name: {
            type: String,
            required: true,
            lowercase: true,
            unique: true
        },
        password: {
            type: String,
            required: true
        },
        admin: {
            type: Boolean,
            required: true
        }
    }
);

/**
 * create user model and export model
 * 
 */
export const User = model('User', UserSchema);